refactor: delete unnecessary files and create 2 new folders
