+++ calendar/src/index.tsx
@@ -1,6 +1,5 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import './index.css';
 import { App } from './App';
 import store from './redux/store';
 import { Provider } from 'react-redux';
